python-backend/requirements.txt
torch>=2.5.0
torchvision>=0.20.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
scipy>=1.11.0
scikit-image>=0.20.0
trimesh>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
svg.path>=6.0